overhaul of db....using user instead of email. fix registration.

+++ ASGShared/Models/Budget.cs
@@ -2,12 +2,12 @@
 {
     public class Budget
     {
-        public decimal Amount { get; set; } = 0;
+        public int Amount { get; set; } = 0;
         public string Currency { get; set; } = "USD"; // Default to USD
 
         public Budget() { }
 
-        public Budget(decimal amount, string currency)
+        public Budget(int amount, string currency)
         {
             Amount = amount;
             Currency = currency;
